break preservator into multiple classes
